Move `javaOpt` to `SharedLaunchOptions`
`jvmOtionFile` should be close to `javaOpt` so I moved it from
`BootstrapSpecificOptions` to `BootstrapOptions`
and changed its group to "Launch"
