+++ modules/cli/src/main/scala/coursier/cli/launch/LaunchOptions.scala
@@ -28,10 +28,6 @@ final case class LaunchOptions(
 
   @Group("Launch")
     fork: Option[Boolean] = None,
-  
-  @Help("Add Java command-line options, when forking")
-  @Value("option")
-    javaOpt: List[String] = Nil,
 
   fetchCacheIKnowWhatImDoing: Option[String] = None,
